Add API to get the supervisor of an application
Add `application:get_supervisor/1` as a public API to get the supervisor
of an application. Especially when debugging or introspecting a running
system it can be useful to programmatically access the supervision tree
of an application. This must start from the root supervisor of the
application, before this commit it was not possible to get the root
supervisor of an application using public API's.

Example programs that could use this functionality:
- A web-based version of `observer`
- A supervision tree designer and visualizer